fix: PodcastController
